requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
scikit-learn>=1.3.0
Flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
joblib>=1.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
pytest>=7.4.0
